+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.15.9] - 2024-03-05
+# [1.15.10] - 2024-03-07
 - Updated available NodeJS lambda versions from 10.x/12.x/14.x to 16.x/18.x/20.x
 - Upgraded default NodeJS lambda version from 14.x to 20.x
+
+# [1.15.9] - 2024-03-04
+- Implemented S3 bucket deployment with configuration for static website hosting
+- Added new resource type `swagger_ui`
 
 # [1.15.8] - 2024-02-27
 - Implemented Cognito User Pools ARNs resolving for OpenAPI specification via the key `x-syndicate-cognito-userpool-names` of the `x-amazon-apigateway-authorizer` extension